requirements-visualization.txt
# Visualization packages for enhanced features
pandas>=1.3.0
plotly>=5.5.0
matplotlib>=3.5.0
numpy>=1.20.0

# Geospatial visualization packages
geopandas>=0.10.2
folium>=0.12.1
streamlit-folium>=0.7.0
branca>=0.5.0
shapely>=1.8.0

# Other utilities
requests>=2.27.1
pyarrow>=7.0.0 